Prometheus instrumentation from VICI client (#8)
This change introduces prometheus metrics collected from a strongswan process
through the vici plugin. It collects statistics by inspecting vici payloads
equivalent the results of swanctl --list-conn and swanctl --list-sas calls.

Package vici is ported from https://github.com/bronze1man/goStrongswanVici and
updated with error handling and removal of fmt.Print calls and a bit simpler
APIs.

Package strongswan implements the actually collection of data and reports to the
PrometheusReporter instance. The PrometheusReporter is then mapping the results
to individual metrics.

Metrics are collected with several labels to make it easy to understand if
configuration etc. is on spot.

    strong_duckling_ike_sa_bytes_in_total{child_sa_name="net-net-0",ike_sa_name="gw-gw",instance="strongswan1:8000",job="strongswan1",local_ip_range="10.101.0.1/32",local_peer_ip="172.22.0.2",remote_ip_range="10.102.0.1/32",remote_peer_ip="172.22.0.3"}

Tracking of states on both the IKE SA and child SAs are not implemented as we do
not know what the values of those fields will be.