Fix stuck command termination
If a signal stops the application it gets stuck when daemon go routines tried to
signal component ending.
This is not the intended behaviour and is fixed with this change.

Only signals and an error on the HTTP listners are able to initiate a shutdown.
As the daemons have no reliable way of indicating that they've stopped, there is
no need for this mechanism.