Remove error state from ClientConn (#27)
Currently method errors are stored in ClientConn.lastError and requires callers
of ClientConn.readResponse to check this field after each call. This is error
prone as it hides the fact that readResponse can fail.

This change removes the ClientConn.lastError field and instead make
ClientConn.readResponse return an error along with the segment. Calling code is
updated accordingly.