Close on already registered vici errors
Currently if strong-duckling fails to read an event request the event handler is
not unregistered as the methods return the request error. This can lead to a
failure state where following stat collections try to register an event handler
but fails as it is already registered as it was not unregistered before.
Strong-duckling is unable to recover from this state.

This change moves unregistration of event handlers into a defered function to
ensure that if registration succeeds we will always attempt to unregister.

RegisterEvent guarantees that it does not keep the handler registered in memory
if any part of the vici registration process fails, which gives us the guarantee
that we will never get into the mentioned bad state again.

Signed-off-by: Bjørn Sørensen <bso@lunar.app>